[AF] Unify and fix meaning and desciption for honorific title strings
A meaning attribute should be specified for the proper translation.
Prior to this change, it was not specified for Desktop string.

Moved the explanation of the title context from meaning to description
attribute on iOS and Android. Previously the string was mistranslated,
since linguists can't see the provided meaning value.

Also use the same meaning on all platfroms to have the same traslation
on all platforms.

Bug: 1141884
Change-Id: Ic20f7e8fa4eed136046e307c27feb7d7b5c81f78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3017982
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Elizabeth Popova <lizapopova@google.com>
Cr-Commit-Position: refs/heads/master@{#901051}
NOKEYCHECK=True
GitOrigin-RevId: 30eb56c663f553062a7b5b050dfee3396c83d1c1