[iOS][MICE] Add a skip limit for the websignin
This patch adds kSigninWebSignDismissalCount preference to count the
number of time the user skips the web sign-in dialog.
A feature param ("consecutive_active_dismissal_limit") is added to have
a max value.
Once the count reaches the max value, the web sign-in dialog is not
shown, and SUPPRESSED_CONSECUTIVE_DISMISSALS metric is recorded.

Also this count is reset to 0 once the user signs in using the web
sign-in.

Fixed: 1218436, 1218430
Change-Id: I85953944a8ce1662a866a5f154599892fc7e2195
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3021236
Auto-Submit: Jérôme <jlebel@chromium.org>
Commit-Queue: Jérôme <jlebel@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901120}
NOKEYCHECK=True
GitOrigin-RevId: e3c8fc519e3d1d0279f6f74f1bcd02b095f908cd