🐻[Pandora] Don't show sync error message as it is deprecated
Previously, no matter if the user click on "Turn on Sync" or
"Don't Sync" on sync screen, the sync error message will always
shown(asking for sync password if have one). It shouldn't show up
if the user doesn't have a intent to sync. Also, the sync engine
won't start if the user haven't tap on "Turn on Sync", so the error
message should be shown in this case.

The original change add checks if the user tap on "Turn on Sync"
to decide whether to show the sync error message. However, since the
Sync error message is deprecated, we just remove it. So, after the
fix, the sync error message won't be shown.

More context:
MICE is deprecating the sync error message(crbug/1151287).
https://docs.google.com/presentation/d/1xqVrulp3OZw-jZuELcNOvgin4gd0svG0XWxfkqgfaMg/edit?resourcekey=0-TzNOA87KdxM48AR2wdeoYg#slide=id.gbb0b810cb2_0_145

Bug: 1225489
Change-Id: Ic153a18d553e3feedab0c9967e6327f2a415f81d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3021697
Reviewed-by: Vincent Boisselle <vincb@google.com>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901533}
NOKEYCHECK=True
GitOrigin-RevId: 5b7126158854fadd99774cc84b4c0a6051332adc