[Fixit] Add utility method to MockConfigurationPolicyProvider.
A utility method is implemented which reduces the amount of boilerplate
code required when using MockConfigurationPolicyProvider.

This is implemented as a helper rather than default behavior (i.e. in
the mock constructor) so that tests which set more specific mock calls
don't produce false positives/negatives.

ON_CALL is used instead of EXPECT_CALL for defining the default mock
method behavior since there may not be any expectation for the mocks to
be called. NiceMock<...> is used to suppress the "Uninteresting mock
function call" warnings.

This cleanup is part of Chrome Fixit 2021.

Bug: 242087
Change-Id: I86f001bf5a41224a1422e52ecb026fa8f16d6482
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3017141
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Igor Ruvinov <igorruvinov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901636}
NOKEYCHECK=True
GitOrigin-RevId: 16185a42bf71818a8494e712850381a66f3ec1b5