Disable tab grid bulk actions buttons based on selected tabs
The share and add to buttons should only be enabled when tabs which are
shareable are selected. Additionally, non-shareable tabs should be
ignored when sharing.

This CL handles both these cases by tracking a list of selectable items
separate from the generic list of all items. This also prevents the need
for any filtering when an action is triggered.

Bug: 1220184, 1225069, 1225289
Change-Id: I904983c73211810273dfc80e354e687d00416d5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3024503
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#902068}
NOKEYCHECK=True
GitOrigin-RevId: c0ad10c0866b21c28890d68b8599ffe9dc74a2dd