[Reland][iOS] Make WebStatePolicyDecider::ShouldAllowRequest asynchronous
This CL updates the implementation of |ShouldAllowRequest| from
returning a decision to instead taking a callback that it calls with a
decision. With these changes, it can support more cases which we need to
wait asynchronous operation finished to determine the final policy
decision.

Bug: 1218389,1226828,1227594
Change-Id: I59f73639fd05afca972e50c504703cbee6c5f4c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3017541
Reviewed-by: Vincent Boisselle <vincb@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Zengfeng Ye <zenye@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#902389}
NOKEYCHECK=True
GitOrigin-RevId: ea38a23e68b05a6860057121444214296e3c95af