[iOS][MICE] Dismiss properly the web sign-in dialog
This patch dismiss the consistency view controller by calling dismiss
on its presenting view controller. This is to make sure it is dismissed
even if there is a view controller still presented on top of the
consistency view controller.

The bug to fix is the following scenario:
 * The sign-in is started
 * The sign-in fails because the identity is not available anymore
 * An alert dialog is presented (about the sign-in failure)
 * The mediator receives the notification that the identity disappeared
 * The mediator sends the notification to the coordinator to disappear
 * The coordinator stops the alert dialog
 * The alert dialog starts to dismiss itself
 * The coordinator starts to dismiss itself

Result: the coordinator dismisses the alert dialog instead of the main
view controller, by calling:
[self.navigationController dismissViewControllerAnimated:completion:]

The fix is to call:
[self.navigationController.presentingViewController
  dismissViewControllerAnimated:completion:]
to be sure to dismiss the main view controller.

This patch also removes the listener to the IdentityManager as soon as
the coordinator starts to shutdown to avoid cookie notifications.

This patch needs EG Tests: crbug.com/1232073

Fixed: 1230008
Change-Id: Ied3f54e90b7a44883326feb752bedd07bf759850
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3042143
Commit-Queue: Jérôme <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904373}
NOKEYCHECK=True
GitOrigin-RevId: 223458427b71b1b4982a517b308d776bef8a7cf3