Backgroun download: Update database for download progress.
On other platforms, the bytes downloaded is persisted in history db,
and write to the background download proto db after download completion.

On iOS, since we don't use history db, update progress should be
directly written to background download proto db.

Also a throttle is used to limit the frequency of db update.

When download is finished, we read the file size and also update the db
with the actual file size.

Bug: 1215315
Change-Id: I19d9a479a67ef21ef9896d4b3c4fed8b56f3b067
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3043448
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904497}
NOKEYCHECK=True
GitOrigin-RevId: cec7cc2e7229d1e662211a40a9efec9f13534c0e