[iOS][MICE] Use bottom sheet if a trait dimension is compact
On landscape for max iPhone, the width of the screen is regular (instead
of compact, for other phones). This patch makes sure that if one trait
collection dimension is compact, the web sign-in view is presented as
bottom sheet instead of a regular dialog like it is on iPad (with
regular size screen).

Fixed: 1231723
Change-Id: I6b6cdc2585abaf452b9bd2b4d6b87e9ef8d9c6d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3042144
Auto-Submit: Jérôme <jlebel@chromium.org>
Commit-Queue: Jérôme <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904697}
NOKEYCHECK=True
GitOrigin-RevId: 9bbfedf7b300dc28e64c010297ec013072867f44