[iOS] Fix text alignment in experimental omnibox
Before, adding the text scrolled the text field to the end, then moving
the cursor back caused the cursor to be at the left side of the text
field with mostly the autocomplete text visible.

Now, the cursor is moved to the beginning, and then moved to the end of
the user's text, so the cursor is mostly at the right side of the text
field and the already-typed text is visible.

Bug: none
Change-Id: Ifcda19213e3f50a400c79cb4afa053f07a3820f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3053700
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#905256}
NOKEYCHECK=True
GitOrigin-RevId: b744426788cd777384ab9daae337e3e1cd4fa366