Create DisableNonHTMLScreenshotOnIOS15 to control PDF screenshots
Create a flag to control the workaround for the iOS bug that cause web
process crashes when taking PDF screenshot.
When enabled (default), the workaround is in place and PDF screenshots
are not taken on iOS15 devices.
When disabled, PDF screenshots are taken normally.

There is no in-app control for the flag.

Bug: 1222426
Change-Id: I3bb5297148e25f9f0de0afc86a1175235aea4da8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3058231
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#906123}
NOKEYCHECK=True
GitOrigin-RevId: b4a23050f58faf305c9bfa06a5c6c16ccfd533a3