Background download: Pass target path directly to BackgroundDownloadTaskHelper.
Currently the logic to generate download file path on iOS lives in
BackgroundDownloadTaskHelper. This is not ideal that the target path
is also needed in BackgroundDownloadServiceImpl.

This CL passes target path to BackgroundDownloadTaskHelper, and updates
the target path in proto db in BackgroundDownloadServiceImpl.

Bug: 1215315
Change-Id: I122aaf1fa9e6d281add6dfbc58e12d098584fab8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3053420
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#906318}
NOKEYCHECK=True
GitOrigin-RevId: a99e431db1c1eaa7d6a21da44033b3de8b3035c9