[iOS] Resize FRE banner pictures manually
In FRE pictures, there is a transparent margin to ensure that the text
is perfectly placed. But when the image was created and got the fill
constraint, the banner area was filled with the current size of the
picture.'UIViewContentModeScaleAspectFill' is not able to reduce the
picture.

The problem was in a smaller screen with a smaller area, the translucent
margin was cropped.

This CL corrects this by resizing pictures manually.

Here are the results on multiple screens:
https://drive.google.com/drive/folders/1zrOP_PRm9NCt2Y5AgX2nBXl2s9mQ38Zz?resourcekey=0-1kYANvuuVCDfssLR1OdM2w&usp=sharing

Fixed: 1220759
Change-Id: Ifda20a3b119f7d637a14e045b38cbe5f9f8d74d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3058235
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Aliona Dangla <alionadangla@google.com>
Cr-Commit-Position: refs/heads/master@{#906675}
NOKEYCHECK=True
GitOrigin-RevId: b849ec3bd52e26dcd7557ceb9efb76f21278128d