[iOS] Fix flake in SafeBrowsingServiceTest.RealTimeSafeAndUnsafePages
This test sometimes crashes because it tries to mark URLs as safe or
unsafe before the Safe Browsing database has been initialized. The
Safe Browsing database uses its own thread, so waiting for the main
thread to be idle does not guarantee that the database is ready.

This CL changes the test so that it first sends an initial query
to the database and waits for a result before marking any URL
as safe or unsafe. Receiving a result for this initial query implies
that the database has been initialized.

Change-Id: I31cb38736400806fec1b742e84b1154a00b3b17e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3061003
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Auto-Submit: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#906843}
NOKEYCHECK=True
GitOrigin-RevId: 257d90f7fb9a4ce3d0bdc40f829651c070cb67a9