[iOS] Fix flakiness for testSigninAddAccountFromWeb
To copy the behavior of ChromeIdentityInteractionManager,
FakeChromeIdentityInteractionManager should not return an identity when
being canceled or when it simulate an error.

Also, this patch clears the identity in
FakeChromeIdentityInteractionManager after being returned, to reset its
state for the next test.

Fixed: 1234458
Change-Id: I4dcbe711db461e6ca72d355e86010a8cda57351c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3060496
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Jérôme <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#906869}
NOKEYCHECK=True
GitOrigin-RevId: 5597c904fe185cd98a1b5111382ca38621fe7412