Add meaning attribute to disambiguate "Name" strings
In some slavic languages, different words for “Name” are used for people
and for inanimate objects which can not be used interchangeably. Added a
meaning attribute to all "Name" strings ("Name of a person" or "Name of
an inanimate object") to create 2 new messages with valid context and
trigger separate translations.

Details: go/name-localization

Added screenshots will be stitched into 2 images (per meaning) and
visible to the localizators for each translation.
IDS_SEARCH_ENGINES_EDITOR_DESCRIPTION_COLUMN is marked not translateable
for now, since it is not visible in the UI and is used only as a column
id in code.

Bug: 1216072
Change-Id: I119d7a9ce7c7e37f40e0562f01081f355f8361cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3001276
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Elizabeth Popova <lizapopova@google.com>
Cr-Commit-Position: refs/heads/master@{#906899}
NOKEYCHECK=True
GitOrigin-RevId: 06a8add71b7be3a84b22196dea51105d1962c8a9