Implement Tab Grid Edit mode Add To Bookmarks
Add multiple selected items to bookmarks from the Tab Grid edit mode.

The BookmarkInteractionController is updated to handle the flow when
bookmarking multiple items by first prompting the user to choose a
destination directory and then saving the bookmarks to that location.

Additionally, the handling of the BookmarksCommand moves to the TabGrid
from the SharingCoordinator in order to use the same bookmarks logic
without code duplication.

Bug: 1196946

Change-Id: Ib4ca4cc3a87a3f2d94b333185f2ea3d5d08bbe33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3045139
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#907201}
NOKEYCHECK=True
GitOrigin-RevId: 715ea46f6b0fd09e2455dfd9311dd3a1a2f5c02c