Always show folder picker when Bookmarking from Tab Grid
The user should always be presented with the folder chooser when
bookmarking from the Tab Grid Edit mode in order to provide a consistent
experience. A single item will still be automatically bookmarked with
the option to Edit as this is the pre-existing behavior before Tab Grid
Edit mode was introduced.

Additionally, ensure that the Edit flow is triggered from the Edit
Bookmark item in the share sheet instead of re-bookmarking the already
existing item.

Fixed: 1235727, 1235728
Change-Id: Icc45e210dbae4379b604c672ca3a3ab32928351c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3065006
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#908564}
NOKEYCHECK=True
GitOrigin-RevId: 1547a5e44db2fc68722998dba3beeb6ddfab641a