[ios] Reading List Messages UI polish
This CL has a few UI polishes. The icons are changed to better sizing.
The text in the modal and below the toggle switch in the Reading List
are modified to use a subtitle font and the correct light gray color.
Finally, the bubble presenter is modified to point up for iPads to
ensure it doesn't clash with the native menu bar and for landscape mode
since the overflow menu is at the top.

Fixed: 1236713, 1231781
Change-Id: Ib554fbfd6ae82dae49a79ae8b210f74aff720c8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3072413
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#908969}
NOKEYCHECK=True
GitOrigin-RevId: c1dc299eff75a89a9901e86c05714ea8353d2b35