[ios] Use the new `signin_error` API in //ios/chrome/browser
Use the `signin_error` API exposed as free function instead of the
old API going through the `SigninErrorProvider` class.

The class `SigninErrorProvider` is not removed yet to prevent breaking
the internal repository when this CL lands. It will be removed in a
followup CL once the override has been deleted.

This is part of a multi-CL change:
  - https://crrev.com/c/3069140
  - https://crrev.com/i/4025853
 => https://crrev.com/c/3067994
  - https://crrev.com/i/4025854
  - https://crrev.com/c/3065652

Bug: 1201182
Change-Id: I7c17ae930e936936f9783f665d882c51ff32bc23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3067994
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Jérôme <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909231}
NOKEYCHECK=True
GitOrigin-RevId: e94f91e96463e1a2b12e22998ad5964f077eb419