[ios] Use the new `signin_resources` API in //ios/chrome/browser
Use the `signin_resources` API exposed as free function instead of
the old API going through the `SigninResourcesProvider` class.

The class `SigninResourcesProvider` is not removed yet to prevent
breaking the internal repository when this CL lands. It will be
removed in a followup CL once the override has been deleted.

Cache the identity's avatar in FakeChromeIdentityService as there
are test checking the cached avatar is indeed cached. Since the
new provider return a new image each time, the test was failing.
Fixed it by adding a cache in FakeChromeIdentityService.

This is part of a multi-CL change:
  - https://crrev.com/c/3067233
  - https://crrev.com/i/4025855
 => https://crrev.com/c/3067234
  - https://crrev.com/i/4025856
  - https://crrev.com/c/3067995

Bug: 1201182
Change-Id: I8a3e412c8b081802cd5ed40b3db7405aa0dcaab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3067234
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Jérôme <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909293}
NOKEYCHECK=True
GitOrigin-RevId: c4d2ec0355ff072e21baeb5eb4c0c75549a3a2d2