[iOS] CPE: Warn user when they are updating an existing password
This CL shows an alert when the user tries to create a new password for
a credential that already exists. If the credential exists in the
UserDefaultsStore that the extension writes to, the extension updates
the credential automatically. If the credential exists in another store, it
the extension creates a new credential in the UserDefaultsStore and
relies on the Chrome-side code to handle resolving the differences.

Bug: 1223966
Change-Id: Ib88086e9701d9b7eeb79a5005fbef3a8b5aed0b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3081821
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Javier Flores <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910540}
NOKEYCHECK=True
GitOrigin-RevId: b0e2b1d8c0e7abb25a6d6cd64488c57019a19152