Change transport_security_persister_path param to be a path to a file.
This CL changes transport_security_persister_path parameter in the
NetworkContextParams to be a full file path rather than a path
to a folder that contains a fixed file called "TransportSecurity".

This is to support the network sandbox as ACLs need to be set on
these files, so the full path must be known by the content layer
rather than being only known by net layer.

This CL contains no functional changes and, in particular, does not
change the path of the transport security persistent data.

However, this CL intentionally changes the name of the parameter
in the network context parameters and the order of the constructor
parameters to ensure all callers update their code to pass a full
file path rather than a path to a directory.

BUG=1173622

Change-Id: Ie135f90b76ef165fa60fbbb295aa5be792d28274
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3082756
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910861}
NOKEYCHECK=True
GitOrigin-RevId: 62f94bd5adf55b3a7eae0c7a14fcf957a3c79d7d