Add another group to the Extended Variations Safe Mode experiment.
After discussions with torne@, among others, it was decided that it
would be helpful to compare the performance of writing synchronously
to disk via PrefService::CommitPendingWriteSynchronously() and writing
to disk via base::WriteFile(). CommitPendingWriteSynchronously() uses
fsync(), which is an expensive operation while WriteFile() does not.

This change modifies the code in several ways to add support for the
new experiment group.
* Plumbs the user data dir file path through to the CleanExitBeacon via
  the MetricsStateManager.
* Implements writing to and reading from a new Variations Safe Mode file
  in clean_exit_beacon.cc.
* Updates the experiment groups in variations_field_trial_creator.cc.

See the description of http://crrev/c/2898323 for more details about
the motivation behind this experiment.

Bug: b/184937096
Change-Id: Ia6d00bb79726913393f86f6e7d51a5f933bde647
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3062519
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#911469}
NOKEYCHECK=True
GitOrigin-RevId: e20a3d498443f9d36f6a849cc5212bba3802fc7c