[iOS] Revert sign-in when interrupting Advanced Settings Sign-in
When the Advanced Settings Sign-in is interrupted (for opening an URL,
or closing Chrome), the sign-in should be reverted, so the user stays
signed out,
But the revert should not happen when opening a Chrome Sync web page,
so this page can be opened.

EG tests for opening a random URL has been updated. EG tests to open
Chrome Sync web page is updated with:
crrev.com/i/4047374

Fixed: 1238746
Change-Id: Ic419039eace727253e089bd65ec91e80597c6750
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3086450
Commit-Queue: Jérôme <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#912158}
NOKEYCHECK=True
GitOrigin-RevId: 854e696e3be965ea03e5bc617f8c19d3216ca6ab