[iOS] Remove ChromeIdentityService::GetAvatarForIdentity() callback
callback cannot be trusted since the avatar might be nil. Plus the
callback is not used anymore in Chrome.
This patch removes it.
Once GetAvatarForIdentity() is called, the caller should expect
Observer::OnProfileUpdate() to be called once the avatar is available.

=> crrev.com/c/3103008
 * crrev.com/i/4066749
 * crrev.com/c/3103367

Bug: 1237022
Change-Id: I52f63cdc89786b2f30ee07574a1e453f0a82c5fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3103008
Commit-Queue: Jérôme <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913563}
NOKEYCHECK=True
GitOrigin-RevId: 85176488f11177001ee275c51e332bc201a62590