Always use MultiStorePasswordSaveManager
PasswordSaveManager has two implementations: PasswordSaveManagerImpl
(the original one; supports only the profile store) and
MultiStorePasswordSaveManager (supports both the profile and the account
store).
Before this CL, which of the two was used depended on the feature flag
kEnablePasswordsAccountStorage.

This CL changes this so that MultiStorePasswordSaveManager is always
used. This is a no-op on platforms where kEnablePasswordsAccountStorage
is enabled (i.e. everything except Android and iOS). On Android and
iOS, this does change the PasswordSaveManager implementation, but in
the absence of an account store, this should make no difference.

Unconditionally using MultiStorePasswordSaveManager required one simple
change to its implementation: Allowing a null account_form_saver.

Bug: 1108738
Change-Id: I0bbbf8fbff382b4f149faa34284df9fe65b9a1d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3097276
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913865}
NOKEYCHECK=True
GitOrigin-RevId: 4282ee52e9a34ef5f04c43d3e5dc9d13c33b7622