[ios] Use the new `url_rewriters` API in //ios/chrome/browser
Use the `url_rewriters` API exposed as free function instead of
the old API going through the `BrowserURLRewriter` class.

The class `BrowserURLRewriter` is not removed yet to prevent
breaking the internal repository when this CL lands. It will be
removed in a followup CL once the override has been deleted.

This is part of a multi-CL change:
  - https://crrev.com/c/3116245
  - https://crrev.com/i/4080488
 => https://crrev.com/c/3116246
  - https://crrev.com/i/4080489
  - https://crrev.com/c/3114292

Bug: 1201182
Change-Id: Ifce9b14e785637cc3b56097523a04aef06ec5b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3116246
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915166}
NOKEYCHECK=True
GitOrigin-RevId: f5c22602c303a7cce07d6a482a51acefebcb59d8