Fix print controller auto dismissal
crrev.com/c/3086223 unveiled an issue when presenting a Print controller
while an activity controller is visible.
In that case, the print controller is dismissed when the activity
controller is dismissed, even if the UIPrintInteractionControllerDelegate
provides a parent view.

Present the print controller after the dismissal to avoid the issue.

Bug: 1240348
Change-Id: Ib31d3172f03397f66755aa29ec99bca423957d15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3113933
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915169}
NOKEYCHECK=True
GitOrigin-RevId: 70f1d230d60e77776690a08a35fc43a4c08660fe