[ios] Enable Discover Feed and Refactored NTP by default
- Deletes Content Suggestion Showcase which is no longer supported
after the discover feed and refactored ntp are enabled by default.

- Creates a Disconnect method to turn off certain services, this is
needed to clean up after tests (and it was probably a bug as well which
we weren't hitting). This new method is needed because we're currently
starting/stopping the Coordinator when showing/hiding the feed, and
disconnect only needs to be called once. This won't be an issue once
we stop doing this.

Bug: 1107599
Change-Id: I00d9d17019751b3fd50043a5175f79697cde6243
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3107707
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915253}
NOKEYCHECK=True
GitOrigin-RevId: dc31c477a4a7b7e5961aea850491d3afe1ce96e6