[omnibox] Start moving parts of OmniboxPopupModel into OmniboxEditModel
This CL moves some logic in the popup model that was primarily accessing
the edit model, into the edit model. Some double model dereferencing
in a using view class is also simplified to access the edit model.
It wasn't necessary to dig into the popup model to get at the
`result` since this was already available via the edit model, and
we want to break any unnecessary dependency on popup model.

Bug: 1208026
Change-Id: I9fdf3c8aac6a0753978e228ec296231bdb76eff0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3125340
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917024}
NOKEYCHECK=True
GitOrigin-RevId: edaaa9cc9edaf84fb2dd38bc3008219f5508887b