[iOS] Fix UpgradeSigninPromo EGTests
UpgradeSigninPromoTestCase was working by luck and not testing
correctly the code.
The issue is that is not possible to setup identities and restart
Chrome and expect those identities to stay.
We were lucky since we were using "test_switches::kSignInAtStartup",
and for other reasons.

We don't need to restart Chrome to test the upgrade sign-in promo. This
patch updates those EGtests and uses the background foreground to
trigger the promo.

This patch also fixes the issue that the promo should not be shown
when the tab grid is active.

Fixed: 1244011
Change-Id: I03ed0c634aebf91ab1022c09c412d6ba3c6f9f2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3123756
Commit-Queue: Jérôme <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917170}
NOKEYCHECK=True
GitOrigin-RevId: 61cd970576fd0082f17daa9c5e9c798501bffee5