[iOS][ForcedSignIn]🏢  Handler for the force sign-in policy
Design Doc:https://docs.google.com/document/d/1ClWqL2BxZu5pj-QgFbVXnQfr5Yss5XX6gQ0n0gHYrc0

Experimental settings screenshot:
https://screenshot.googleplex.com/4UHgsv7VsFk9KAe

Sets the forced sign-in policy pref when the switch allows it.

Add a switch to control the feature until it is ready to be released
(the policy isn't applied if the switch is OFF), hence admins cannot
unexpectedly enable the policy.

Add a new experimental setting for Enterprise that allows setting a
value for the sign-in policy. We introduce a "Default" value to
support the case where the policy isn't set because a pref with a value
of 0 has the effect of disabling sign-in. Default values other than 0
aren't supported for integers in the experimental settings.

Add a new util method that determines whether the forced-sign-in policy
applies.

Bug: 1237165
Change-Id: Ia6f5f045e11c19e38213d6e18a820ef16c7ac1c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076380
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Jérôme <jlebel@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917288}
NOKEYCHECK=True
GitOrigin-RevId: 19200bca224cb3aa8d8d7622a6a4a9819727d360