[ios] Use the new `voice_search` API in //ios/chrome/browser
Use the `voice_search` API exposed as free function instead of
the old API going through the `VoiceSearchProvider` class.

The class `VoiceSearchProvider` is not removed yet to prevent
breaking the internal repository when this CL lands. It will be
removed in a followup CL once the override has been deleted.

This is part of a multi-CL change:
  - https://crrev.com/c/3133123
  - https://crrev.com/i/4095470
 => https://crrev.com/c/3132986
  - https://crrev.com/i/4095471
  - https://crrev.com/c/3133083

Bug: 1201182
Change-Id: I5f79cddf79adbe4172f52e03aede26b78d0ad99e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3132986
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917587}
NOKEYCHECK=True
GitOrigin-RevId: d9dd5e7b5e2c4d6c011f1fd06a885c6f0f06d120