[iOS] CPE: Allow logging of arbitrary histograms in extensions
The histograms are logged into a dictionary where the keys are composed
of the histogram name and bucket number, and the values are the number
of times that histogram bucket has occurred in the extension.

Then the metrics mediator emits those events when Chrome returns to the
foreground.

Bug: 1224964
Change-Id: I01fc3aae3d0e426e75dc232388dc674cd78e3b6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3124817
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Javier Flores <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919063}
NOKEYCHECK=True
GitOrigin-RevId: 51dc1e38518a42a14cc126e8b1081a5810bab94c