[iOS] Fix PasswordSpecFetcher on ASan builds
When PasswordSpecFetcher receives data as NSData*, it uses the
underlying data as a const char* that is passed to
base::Base64Decode. However, this data is not null-terminated,
so Base64Decode will keep reading past the end of this array until
a null character happens to be reached. This causes
PasswordSpecFetcherTest.DefaultSpecInvalidFetch to fail on the
ios-asan bot.

This CL safely converts the received data into a StringPiece with
a specific size, fixing the ASan failure.

Bug: 1248514

Change-Id: I83d441d00f58b9ca8f53990c31ef448d7b602608
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3152561
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#920320}
NOKEYCHECK=True
GitOrigin-RevId: 521a165cc86cf091072e4ed569b1dcdd78b92077