[iOS] Disable drag-and-drop on omnibox on iOS 15 iPhone temporarily
iOS 15 adds Drag and Drop support to iPhones. This causes the long-press
recognizer for showing the copy/paste menu to not appear until the user
lifts their finger. The long-term solution is to move to the new
UIContextMenu API, but for now, disable Drag from the omnibox on iOS 15
iPhones.

Bug: 1247668
Change-Id: If1246c29c4f3e5f760fdad3f7dcf2938bcd21bd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3157607
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#920832}
NOKEYCHECK=True
GitOrigin-RevId: b5fa3252e1f3157fdcc6791c7976b27e54f6c376