[iOS] Restore full window opening behavior for native menus.
The native context menu implementation for web content doesn't match
the old (non-native) behavior for "open in new window". The new code
doesn't pass referrer or incognito status in the activity used to open
load the link in the new window.

Not passing incognito status is a user-visible regression, since opening
tabs from an incognito page in a new window will open non-incognito tabs.

(Not passing the referrer is also a regression, albeit one with far less
obvious effects).

This CL addresses this by adding an action factory method that can
open a window with an arbitrary user activity, and then the BVC uses
the existing activity creation function that accepts referrer and
incognito status. Since other code uses the preexisting method that
just takes a URL and an origin, it's left unchanged.

An EG test that checks for this regression will need to be added, but
since this CL will be cherrypicked into M94 and M95, I wanted to keep
it as small as possible; the EG test will land on trunk from a follow-up
CL.

Bug: 1246019
Change-Id: I488b6bf1d86f75084999253610ee0de23d6ca46e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3157891
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921163}
NOKEYCHECK=True
GitOrigin-RevId: b7e9657badc28167eb0faf55cfc30531e28021eb