Add WebTestWithWebState helpers for JavaScriptFeatures
When executing JavaScript, tests will often need to do so in the same
content world as a particular JavaScriptFeature that is being tested.
ExecuteJavaScriptForFeature and CallJavaScriptFunctionForFeature allow
for a test to execute such JavaScript without needing to interact with
other intermediary objects.

Bug: 1206328
Change-Id: I911d99b1440a58250045e81bd8360ebd6b982056
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3160803
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921650}
NOKEYCHECK=True
GitOrigin-RevId: 297bf2e0894b0d7435ea82cc37a414d811ec6d62