[Metrics Variations] Leverage BuildState for browser version information
Updates the Chrome Variations service client to leverage the BuildState
class for obtaining installed browser version information. With this
change, we can now remove a static UpgradeDetectorImpl function that was
exclusively being used by the Chrome variations service client.

For this to be possible the VariationsServiceClient's version for
simulation function is now synchronous and will run on the UI thread
to leverage the BuildState's version data.

Bug: 1249640
Change-Id: I28f854fb1cc2ff7b949841ec9903bff38afd30d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3161376
Commit-Queue: Roman Arora <romanarora@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#922255}
NOKEYCHECK=True
GitOrigin-RevId: c33192baeeeee0ea7c2b591957219cd6baf8fe13