[iOS] [Apfelstrudel] Use Refactored NTP when feed is not visible
Allows the refactored NTP to be used in all cases (was previously only
used when the feed was visible). When there is no Discover feed, an
empty scroll view is used as the primary NTP scroll view. This scroll
view maintains the correct minimum height to allow for the omnibox to be
focused.

Bug: 1250724
Change-Id: I261d4d244d767a5835f52e4872f055d61bc00127
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3169112
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923170}
NOKEYCHECK=True
GitOrigin-RevId: 57bad25ae8d0d423983c5006ca1bf493f97f4182