[iOS] Remove baseVC from init of image utils
This CL passes the baseViewController in the method calls of ImageSaver
and ImageCopier, instead of passing it in the init.
This will allow the ContextMenu code to be extracted from BVC.

It also fixes a bug where -canOpenURL: should be called from the main
thread.

Bug: 1250689
Change-Id: Iecf6afbc1968e20377b1d765bc5d0008ccc60fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3171057
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923367}
NOKEYCHECK=True
GitOrigin-RevId: 04c548a7227128a50da70bc20f2445fc824017ae