[iOS][ForcedSignIn]🏢Dont allow intents dismissing forced sign-in prompt
Intents cannot dismiss the forced sign-in prompt because sign-in is
required in that context.

--
There are 3 scenarios where the intents can be handled that needs
particular handling when forced sign-in policy is active:
  (1) At startup when the policy is enabled
  (2) When the browser is active and showing the forced sign-in prompt.
  (3) When transitioning the ownership of the forced sign-in prompt to
      another scene.

Checking that `IsSigninForcedByPolicy() && ![self isSignedIn]` is false
will make sure that the intent isn't handled when
AppState.signinInProgress == NO during the browser startup OR the
transition between 2 scenes (moments (1) and (3)).

For (2), we need to handle the intent when |signinCoordinator| is set
to nil, which means that the sign-in prompt is dismissed.

See http://shortn/_ZvhvM0oDEr for more details.
--

Bug: 1238317
Change-Id: Iff51cf07879b2729f57b5528105c6e286145c8fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3152845
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#923433}
NOKEYCHECK=True
GitOrigin-RevId: 5b8d4b66c675258b2eb2f03d5e2230dbc38a2eac