Update the saved password warning dialog.
Update the saved password warning dialog to be
more aligned with the other check passwords
warning dialog. Because the string no longer
contains the domains, this CL removes all
references to bolding parts of the string
which used to be the domains.

Screenshot:
http://screen/8DJeHij8EuxSfyF.png

Bug: 1248246
Change-Id: Ia662116353cecab7ce16f6780feff92d2e54c132
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3153716
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923497}
NOKEYCHECK=True
GitOrigin-RevId: 03a25827ef0a5a3fd41f1391669cf1baa995c452